requirements.txt
Django==2.1.2
djangorestframework==3.9.0
django-rest-framework-mongoengine==3.3.1
mongoengine==0.15.3
requests==2.20.0
beautifulsoup4==4.6.3